refactor(mall): 重构mall后调整适配所有接口

+++ services/rpc/mall/internal/logic/orderservice/list_orders_logic.go
@@ -30,7 +30,7 @@ func (l *ListOrdersLogic) ListOrders(in *pb.ListOrdersReq) (*pb.ListOrdersResp, 
 		Page:   int(in.Page),
 		Size:   int(in.PageSize),
 		UserId: in.UserId,
-		Status: int64(in.Status),
+		Status: int(in.Status),
 	}
 	if req.Status < 0 {
 		req.Status = -1
