+++ services/rpc/mall/model/mall_orders/mall_orders_model.go
@@ -18,9 +18,9 @@ type (
 		// InsertTx 在事务中插入订单。
 		InsertTx(tx *gorm.DB, data *MallOrders) error
 		// UpdateStatusTx 使用乐观锁条件更新订单状态，返回是否实际更新成功。
-		UpdateStatusTx(tx *gorm.DB, id, userID string, fromStatus, toStatus int64, now time.Time) (bool, error)
+		UpdateStatusTx(tx *gorm.DB, id, userID string, fromStatus, toStatus int, now time.Time) (bool, error)
 		// MarkPaidTx 使用乐观锁将订单从 fromStatus 标记为已支付，并写入支付方式与支付时间，返回是否实际更新成功。
-		MarkPaidTx(tx *gorm.DB, id, userID string, fromStatus int64, payType string, now time.Time) (bool, error)
+		MarkPaidTx(tx *gorm.DB, id, userID string, fromStatus int, payType string, now time.Time) (bool, error)
 	}
 
 	customMallOrdersModel struct {
@@ -48,7 +48,7 @@ func (m *customMallOrdersModel) InsertTx(tx *gorm.DB, data *MallOrders) error {
 }
 
 // UpdateStatusTx 使用乐观锁条件更新订单状态，返回是否实际更新成功。
-func (m *customMallOrdersModel) UpdateStatusTx(tx *gorm.DB, id, userID string, fromStatus, toStatus int64, now time.Time) (bool, error) {
+func (m *customMallOrdersModel) UpdateStatusTx(tx *gorm.DB, id, userID string, fromStatus, toStatus int, now time.Time) (bool, error) {
 	result := tx.Model(&MallOrders{}).
 		Where("id = ? AND user_id = ? AND status = ?", id, userID, fromStatus).
 		Updates(map[string]any{
@@ -62,7 +62,7 @@ func (m *customMallOrdersModel) UpdateStatusTx(tx *gorm.DB, id, userID string, f
 }
 
 // MarkPaidTx 使用乐观锁将订单从 fromStatus 标记为已支付，并写入支付方式与支付时间，返回是否实际更新成功。
-func (m *customMallOrdersModel) MarkPaidTx(tx *gorm.DB, id, userID string, fromStatus int64, payType string, now time.Time) (bool, error) {
+func (m *customMallOrdersModel) MarkPaidTx(tx *gorm.DB, id, userID string, fromStatus int, payType string, now time.Time) (bool, error) {
 	result := tx.Model(&MallOrders{}).
 		Where("id = ? AND user_id = ? AND status = ?", id, userID, fromStatus).
 		Updates(map[string]any{
